requirements.txt
streamlit>=1.40.0
yfinance>=0.2.40
plotly>=5.22.0
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.1
groq>=0.9.0
google-genai>=1.0.0
openai>=1.30.0
